+++ odoo/local-src/stock_delivery_note/models/stock_picking.py
@@ -60,31 +60,50 @@ class StockPicking(models.Model):
         1: Id (name of picking); email customer
         2: name customer; street customer; zip + city; country
         Next lines are the details of what is send one line by stock moves:
-            Product esb_ref
+            Product esb_ref (default_code)
             Product name
             Product qty
             Net price without VAT
             Crude price without VAT
-            Vat rate ?
+            Vat rate
             Lot ids
             Use dates
-            Suite name ?
+            Suite name
 
         """
+        def format_number(number, fractional_size=None):
+            """Format a number to a string.
+
+            The number is formated separating the decimal and fractional part
+            with a comma. With between 1 and 3 number after the comma.
+            """
+            if fractional_size == 1:
+                formater = '{:.1f}'
+            elif fractional_size == 2:
+                formater = '{:.2f}'
+            elif fractional_size == 3:
+                formater = '{:.3f}'
+            else:
+                formater = '{}'
+            s = formater.format(number)
+            return ','.join(s.split('.'))
+
         self.ensure_one()
         lines = []
         partner = self.partner_id
+        # The two header lines
         lines.append([
-            self.name,
-            partner.email if partner.email else '',
+            self.id,
+            partner.email or '',
             ])
         lines.append([
-            partner.name if partner.name else '',
-            partner.street if partner.street else '',
-            (partner.zip + ' ' if partner.zip else ''
-             + partner.city if partner.city else ''),
-            partner.country_id.name if partner.country_id else '',
+            u'{} {}'.format(partner.title.shortcut or '',
+                            partner.name or '').strip(),
+            partner.street or '',
+            u'{} {}'.format(partner.zip or '', partner.city or '').strip(),
+            partner.country_id.name or '',
             ])
+        # The product lines
         for move in self.move_lines:
             product = move.product_id
             sol = move.procurement_id.sale_line_id
@@ -94,18 +113,20 @@ class StockPicking(models.Model):
             use_date = [ld[-2:] + ld[4:8] + ld[:4] for ld in use_date]
             use_date = '/'.join(use_date)
             lines.append([
-                product.default_code if product.default_code else '',
+                product.default_code or '',
                 product.name,
-                str(sum(quants.mapped('qty'))),
+                # Quantity computed from the quants
+                format_number(sum(quants.mapped('qty')), 3),
                 #  Net HTVA price
-                str(sol.price_reduce),
+                format_number(sol.price_reduce, 2),
                 #  Brut HTVA price
-                str(sol.price_unit),
-                #  VAT rate ? First tax, all taxes, waiting for an answer
-                str(sol.tax_id[0].amount) if len(sol.tax_id) else '0',
-                ' / '.join([str(l.id) for l in quants.mapped('lot_id')]),
+                format_number(sol.price_unit, 2),
+                #  VAT rate, yes only the first one if present
+                format_number(sol.tax_id[0].amount if sol.tax_id else 0, 1),
+                # Lots name
+                '/'.join(quants.mapped('lot_id.name')),
                 use_date,
-                'no suite',
+                sol.order_id.suite_name or '',
                 ]
             )
             return lines