+++ HISTORY.rst
@@ -19,6 +19,8 @@ Unreleased
 
 **Bugfixes**
 
+* ALCN-864: Fixing the csv document generated as a delivery note by improving float number format and adding no suite, lot name, tax amount.
+
 **Build**
 
 * ALCN-1394: Disable the module account_sepa (enterprise) and install the module l10n_be_iso20022_pain